requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2
Pillow
django-filter